# Performance improvements to SELECT API on certain query operations