… pkg/s3select/helpers.go
… pkg/s3select/select.go
… pkg/s3select/select_test.go
… vendor/github.com/json-iterator/go/Gopkg.lock
… vendor/github.com/json-iterator/go/Gopkg.toml
… vendor/github.com/json-iterator/go/LICENSE
… vendor/github.com/json-iterator/go/README.md
… vendor/github.com/json-iterator/go/adapter.go
… vendor/github.com/json-iterator/go/any.go
… vendor/github.com/json-iterator/go/any_array.go
… vendor/github.com/json-iterator/go/any_bool.go
… vendor/github.com/json-iterator/go/any_float.go
… vendor/github.com/json-iterator/go/any_int32.go
… vendor/github.com/json-iterator/go/any_int64.go
… vendor/github.com/json-iterator/go/any_invalid.go
… vendor/github.com/json-iterator/go/any_nil.go
… vendor/github.com/json-iterator/go/any_number.go
… vendor/github.com/json-iterator/go/any_object.go
… vendor/github.com/json-iterator/go/any_str.go
… vendor/github.com/json-iterator/go/any_uint32.go
… vendor/github.com/json-iterator/go/any_uint64.go
… vendor/github.com/json-iterator/go/build.sh
… vendor/github.com/json-iterator/go/config.go
… vendor/github.com/json-iterator/go/fuzzy_mode_convert_table.md
… vendor/github.com/json-iterator/go/iter.go
… vendor/github.com/json-iterator/go/iter_array.go
… vendor/github.com/json-iterator/go/iter_float.go
… vendor/github.com/json-iterator/go/iter_int.go
… vendor/github.com/json-iterator/go/iter_object.go
… vendor/github.com/json-iterator/go/iter_skip.go