@harshavardhana 2018-11-02
Some benchmarking results using `go test -bench .`

```
    benchcmp old new
    benchmark                        old ns/op       new ns/op       delta
    BenchmarkSQLAggregate_100K-4     551213          259782          -52.87%
    BenchmarkSQLAggregate_1M-4       6981901985      2432413729      -65.16%
    BenchmarkSQLAggregate_2M-4       13511978488     4536903552      -66.42%
    BenchmarkSQLAggregate_10M-4      68427084908     23266283336     -66.00%
    
    benchmark                        old allocs     new allocs     delta
    BenchmarkSQLAggregate_100K-4     2366           485            -79.50%
    BenchmarkSQLAggregate_1M-4       47455492       21462860       -54.77%
    BenchmarkSQLAggregate_2M-4       95163637       43110771       -54.70%
    BenchmarkSQLAggregate_10M-4      476959550      216906510      -54.52%
    
    benchmark                        old bytes       new bytes      delta
    BenchmarkSQLAggregate_100K-4     1233079         1086024        -11.93%
    BenchmarkSQLAggregate_1M-4       2607984120      557038536      -78.64%
    BenchmarkSQLAggregate_2M-4       5254103616      1128149168     -78.53%
    BenchmarkSQLAggregate_10M-4      26443524872     5722715992     -78.36%
    ```