@harshavardhana 2018-11-03
Comparison with [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat)

```
~ benchstat old new
name                 old time/op    new time/op    delta
SQLAggregate_100K-4    33.9ms ± 0%     0.3ms ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_1M-4       7.46s ± 0%     2.36s ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_2M-4       14.2s ± 0%      4.7s ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_10M-4      79.8s ± 0%     23.5s ± 0%   ~     (p=1.000 n=1+1)

name                 old alloc/op   new alloc/op   delta
SQLAggregate_100K-4    10.0MB ± 0%     1.1MB ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_1M-4      2.72GB ± 0%    0.61GB ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_2M-4      5.45GB ± 0%    1.22GB ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_10M-4     27.2GB ± 0%     6.1GB ± 0%   ~     (p=1.000 n=1+1)

name                 old allocs/op  new allocs/op  delta
SQLAggregate_100K-4      157k ± 0%        0k ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_1M-4       48.2M ± 0%     22.0M ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_2M-4       96.5M ± 0%     44.0M ± 0%   ~     (p=1.000 n=1+1)
SQLAggregate_10M-4       482M ± 0%      221M ± 0%   ~     (p=1.000 n=1+1)
```